refactor: 버튼 공통 컴포넌트 타입 전역적으로 한번에 관리하도록 수정
